+++ packages/styles/src/packages/atomic/iconButton/rarui-iconButton.css.ts
@@ -19,7 +19,7 @@ const iconButton = recipe({
     fontWeight: varsThemeBase.fontWeight.semiBold,
     lineHeight: varsThemeBase.lineHeight.button.l,
     fontSize: varsThemeBase.fontSize.button.l,
-    fontFamily: varsThemeBase.fontFamily.inter,
+    fontFamily: varsThemeBase.fontFamily.base,
     borderRadius: varsThemeBase.shape.border.radius["2xs"],
     borderWidth: varsThemeBase.shape.border.width[1],
     borderStyle: "solid",
@@ -115,7 +115,7 @@ const iconButton = recipe({
       solid: {
         selectors: {
           "&:hover:after": {
-            backgroundColor: varsThemeBase.colors.surface["on-brand-hover"],
+            backgroundColor: varsThemeBase.colors.surface.hover,
           },
           "&:active:after": {
             borderColor: varsThemeBase.colors.surface.press,
@@ -259,8 +259,17 @@ const iconButton = recipe({
         ...buttonVariants.text.brand,
         backgroundColor: varsThemeBase.colors.surface["brand-subdued"],
         selectors: {
+          "&:hover": {
+            backgroundColor: varsThemeBase.colors.surface["brand-subdued"],
+          },
+          "&:active": {
+            backgroundColor: varsThemeBase.colors.surface["brand-subdued"],
+          },
           "&:hover:after": {
             backgroundColor: varsThemeBase.colors.surface["brand-hover"],
+          },
+          "&:active:after": {
+            backgroundColor: varsThemeBase.colors.surface["brand-press"],
           },
         },
       },
@@ -274,8 +283,17 @@ const iconButton = recipe({
         ...buttonVariants.text.danger,
         backgroundColor: varsThemeBase.colors.surface["error-subdued"],
         selectors: {
+          "&:hover": {
+            backgroundColor: varsThemeBase.colors.surface["error-subdued"],
+          },
+          "&:active": {
+            backgroundColor: varsThemeBase.colors.surface["error-subdued"],
+          },
           "&:hover:after": {
             backgroundColor: varsThemeBase.colors.surface["error-hover"],
+          },
+          "&:active:after": {
+            backgroundColor: varsThemeBase.colors.surface["error-press"],
           },
         },
       },
@@ -289,8 +307,17 @@ const iconButton = recipe({
         ...buttonVariants.text.success,
         backgroundColor: varsThemeBase.colors.surface["success-subdued"],
         selectors: {
+          "&:hover": {
+            backgroundColor: varsThemeBase.colors.surface["success-subdued"],
+          },
+          "&:active": {
+            backgroundColor: varsThemeBase.colors.surface["success-subdued"],
+          },
           "&:hover:after": {
             backgroundColor: varsThemeBase.colors.surface["success-hover"],
+          },
+          "&:active:after": {
+            backgroundColor: varsThemeBase.colors.surface["success-press"],
           },
         },
       },
@@ -304,8 +331,17 @@ const iconButton = recipe({
         ...buttonVariants.text.warning,
         backgroundColor: varsThemeBase.colors.surface["warning-subdued"],
         selectors: {
+          "&:hover": {
+            backgroundColor: varsThemeBase.colors.surface["warning-subdued"],
+          },
+          "&:active": {
+            backgroundColor: varsThemeBase.colors.surface["warning-subdued"],
+          },
           "&:hover:after": {
             backgroundColor: varsThemeBase.colors.surface["warning-hover"],
+          },
+          "&:active:after": {
+            backgroundColor: varsThemeBase.colors.surface["warning-press"],
           },
         },
       },
@@ -319,8 +355,17 @@ const iconButton = recipe({
         ...buttonVariants.text.neutral,
         backgroundColor: varsThemeBase.colors.surface.secondary,
         selectors: {
+          "&:hover": {
+            backgroundColor: varsThemeBase.colors.surface.secondary,
+          },
+          "&:active": {
+            backgroundColor: varsThemeBase.colors.surface.secondary,
+          },
           "&:hover:after": {
             backgroundColor: varsThemeBase.colors.surface.hover,
+          },
+          "&:active:after": {
+            backgroundColor: varsThemeBase.colors.surface.press,
           },
         },
       },
@@ -334,8 +379,17 @@ const iconButton = recipe({
         ...buttonVariants.text.inverted,
         backgroundColor: varsThemeBase.colors.surface["invert-secondary"],
         selectors: {
+          "&:hover": {
+            backgroundColor: varsThemeBase.colors.surface["invert-secondary"],
+          },
+          "&:active": {
+            backgroundColor: varsThemeBase.colors.surface["invert-secondary"],
+          },
           "&:hover:after": {
             backgroundColor: varsThemeBase.colors.surface["invert-hover"],
+          },
+          "&:active:after": {
+            backgroundColor: varsThemeBase.colors.surface["invert-press"],
           },
         },
       },